refactor: split usb.h into usb.c + usb.h pair
Move all USB implementation from usb.h into usb.c. The header
retains only the include of drivers.h.
